// server/src/product/Dtos/addPoduct.dto.ts
export class addProductDto {
  productName: string;
  productDescription: string;
  productPrice: number;
  productStock: number;
  productRating: number;
  productOrigin: string;
  productSeason: string;
  productAbout: string;
  productPicture: string;
  productSecondPicture: string;
  productThirdPicture: string;
  productFourthPicure: string;
}